Markup: Break up fetchArrivals.ts into 3 files - tflApiService.ts that calls the TfL API, postcodeApiService.ts that calls the postcodes API, and fetchArrivals that takes the responses from these 2 files and implements the business logic
